requirements.txt
streamlit==1.32.0
llama-cpp-python==0.2.27
python-dotenv==1.0.1 